move permit lists to PermitType.ts
